+++ app/core/eqemu-data-service-client.js
@@ -87,5 +87,12 @@ module.exports = {
     }
 
     return response.replace('Worldserver Uptime: ', '').trim();
-  }
+  },
+
+  /**
+   * @returns {Promise<*>}
+   */
+  messageWorld: async function (message) {
+    telnetService.execWorld('emote world 15 ' + message);
+  },
 };