+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2020-1-7
+* Implement timed restarts
+* Implement timed restart cancelling
+
 ## [1.0.5] - 2020-1-6
 * Launcher refactoring, shared memory on bootup support
 * Update server processes on dashbaord in realtime